test: add ingress, robots, tunnel edge case tests
Comprehensive testing of edge cases in ingress tunnel sessions, robots.txt
compliance, and tunnel protocol handling to ensure production readiness
for ADR-060 ingress endpoints.

+++ internal/handlers/ingresssessions.go
@@ -52,7 +52,7 @@ func (a *API) ListIngressTunnelSessions(w http.ResponseWriter, r *http.Request, 
 		BeforeID:   before,
 		EndpointID: endpointID,
 		ActiveOnly: params.Active == nil || *params.Active,
-		Limit:      int32(limit) + 1,
+		Limit:      limit + 1,
 	})
 	if err != nil {
 		a.internalError(w, r, "list ingress sessions", err)
